Extract heatmap calculator to its own PORO
