+++ app/controllers/api/v1/analyst/heatmap_controller.rb
@@ -2,7 +2,8 @@ class Api::V1::Analyst::HeatmapController < Api::ApiController
   respond_to :json
 
   def show
-    respond_with CityAnalyst.new(params[:keywords], params[:city])
-                            .heatmap_of(params[:neighborhood])
+    respond_with HeatmapCalculator.new(params[:keywords],
+                                       params[:city],
+                                       params[:neighborhood]).run
   end
 end
